requirements.txt
absl-py==0.9.0
actionlib==1.12.0
angles==1.9.12
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
bondpy==1.8.3
Brlapi==0.6.6
camera-calibration==1.14.0
camera-calibration-parsers==1.11.13
catkin==0.7.20
catkin-pkg==0.4.16
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
controller-manager==0.17.0
controller-manager-msgs==0.17.0
cupshelpers==1.0
cv-bridge==1.13.0
defer==1.0.6
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distro==1.4.0
distro-info===0.18ubuntu0.18.04.1
dlib==19.19.0
docutils==0.16
dynamic-reconfigure==1.6.0
gast==0.3.3
gazebo-plugins==2.8.6
gazebo-ros==2.8.6
gencpp==0.6.2
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.12
gennodejs==2.0.1
genpy==0.6.9
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
httplib2==0.9.2
idna==2.6
image-geometry==1.13.0
imutils==0.5.3
interactive-markers==1.11.4
isort==4.3.21
joint-state-publisher==1.12.14
kdl-parser-py==1.13.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
laser-geometry==1.6.4
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.1
MarkupSafe==1.0
mccabe==0.6.1
message-filters==1.14.3
mtcnn==0.1.0
mysqlclient==1.4.6
netifaces==0.10.4
numpy==1.18.1
oauth==1.0.1
olefile==0.45.1
onnx==1.6.0
onnx-tf==1.5.0
onnxruntime==1.1.2
opencv-python==4.2.0.32
paho-mqtt==1.5.0
pexpect==4.2.1
Pillow==6.2.0
protobuf==3.11.3
pycairo==1.16.2
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-qt-binding==0.4.0
pytz==2019.3
PyYAML==5.3
qt-dotgraph==0.4.0
qt-gui==0.4.0
qt-gui-cpp==0.4.0
qt-gui-py-common==0.4.0
reportlab==3.4.0
requests-unixsocket==0.1.5
resource-retriever==1.12.6
rosbag==1.14.3
rosboost-cfg==1.14.7
rosclean==1.14.7
roscreate==1.14.7
rosgraph==1.14.3
roslaunch==1.14.3
roslib==1.14.7
roslint==0.11.2
roslz4==1.14.3
rosmake==1.14.7
rosmaster==1.14.3
rosmsg==1.14.3
rosnode==1.14.3
rosparam==1.14.3
rospkg==1.2.3
rospy==1.14.3
rosservice==1.14.3
rostest==1.14.3
rostopic==1.14.3
rosunit==1.14.7
roswtf==1.14.3
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.14
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.7
scipy==1.4.1
SecretStorage==2.3.1
sensor-msgs==1.12.7
simplejson==3.13.2
six==1.14.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
sqlparse==0.3.1
system-service==0.3
systemd-python==234
tensorflow==1.15.2
termcolor==1.1.0
tf==1.12.0
tf-conversions==1.12.0
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.2
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.0
wrapt==1.12.0
xacro==1.13.5
xkit==0.0.0
zope.interface==4.3.2
